backend/requirements-windows.txt
# Windows-compatible requirements for Mai-Guru Platform
# Use this file if you encounter build issues with the main requirements.txt

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0

# Database - PyMySQL for Windows compatibility
PyMySQL==1.1.0
django-environ==0.11.2

# AI & Web Scraping
openai==1.3.0
beautifulsoup4==4.12.2
requests==2.31.0

# Payment Gateways
paypalrestsdk==1.13.3
requests-oauthlib==1.3.1

# File Handling - Pre-compiled wheels for Windows
Pillow==10.0.1

# Email & Notifications
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0

# Security
cryptography==41.0.7

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
